requirements.txt
bleach==1.4.1
dj-static==0.0.6
Django==1.8
django-aloha-edit==0.3.3
django-ckeditor==4.4.8
django-tinymce==1.5.3
gunicorn==19.3.0
html5lib==0.99999
lxml==3.4.4
Pillow==2.8.2
psycopg2==2.6
six==1.9.0
static3==0.5.1
